+++ docs/demo/slider/slider-input.jsx
@@ -63,7 +63,7 @@ class Demo extends React.Component {
     return (
       <Row>
         <Col span={22}>
-          <Slider defaultValue={10} onChange={(value)=>this.handleInput(value)} value={this.state.value}  max={90}  min={10}  ifShowRangeLabel={true}/>
+          <Slider defaultValue={10} onChange={(value)=>this.handleInput(value)} value={this.state.value}  max={90}  min={10}  showRangeLabel={true}/>
         </Col>
         <Col span={2}>
           <Input    
@@ -153,7 +153,7 @@ class Demo extends React.Component {
               <Slider
                 vertical
                 defaultValue={30}
-                ifShowRangeLabel={true}
+                showRangeLabel={true}
                 vertical
                 value={value}
                 min={min}